Move coverage lifter into the coverage phase
